+++ Change Log.md
@@ -1,2 +1,18 @@
-v 1.0.1
+v 0.0.5
+- Added tags and adjusted tags to meet itunes/spotify rss requirements
+- Added gui for new tags
+
+v 0.0.4
+- added email address to pod data for itunes:owner
+- fixed bug with txt data at channel level
+
+v 0.0.3
+- fixed issues with enclosure/alternate enclosure/ audio enclosure
+- added better diagnostic messaging to debug further enclosure issues
+
+v 0.0.2
+- added 301 redirect capability to custom RSS
+- disabled podcast guid to prevent duplication.
+
+v 0.0.1
 - Strippled lightning support out to cread podcasting fork